requirements.txt
pandas
openpyxl
supabase
python-dotenv
psycopg2-binary